split PR

+++ catalogs/catalog-lakehouse-iceberg/src/test/java/org/apache/gravitino/catalog/lakehouse/iceberg/TestIcebergTable.java
@@ -535,7 +535,9 @@ public class TestIcebergTable {
       if (icebergCatalogOperations.tableExists(renamedIdentifier)) {
         icebergCatalogOperations.dropTable(renamedIdentifier);
       }
-      icebergCatalogOperations.dropSchema(targetSchemaIdent, false);
+      if (icebergCatalogOperations.schemaExists(targetSchemaIdent)) {
+        icebergCatalogOperations.dropSchema(targetSchemaIdent, false);
+      }
     }
   }
